iot/firmware/requirements.txt
paho-mqtt==1.6.1
dataclasses==0.8; python_version < '3.7'
typing-extensions==4.8.0
